Improve stub generation to always use correct return type.

+++ source/syscall_stubs.c
@@ -5,7 +5,7 @@
 // anhttpSocket
 anhttpSocketArgs_t anhttpSocketArgs[64];
 int anhttpSocketArgsCount = 0;
-int anhttpSocketReturn = 0;
+int anhttpSocketReturn = (int)0;
 int anhttpSocket(int domain, int type, int protocol) {
     if (anhttpSocketArgsCount == 64) assert(0);
     anhttpSocketArgs[anhttpSocketArgsCount].domain = domain;
@@ -18,7 +18,7 @@ int anhttpSocket(int domain, int type, int protocol) {
 // anhttpListen
 anhttpListenArgs_t anhttpListenArgs[64];
 int anhttpListenArgsCount = 0;
-int anhttpListenReturn = 0;
+int anhttpListenReturn = (int)0;
 int anhttpListen(int socket, int backlog) {
     if (anhttpListenArgsCount == 64) assert(0);
     anhttpListenArgs[anhttpListenArgsCount].socket = socket;
@@ -30,7 +30,7 @@ int anhttpListen(int socket, int backlog) {
 // anhttpClose
 anhttpCloseArgs_t anhttpCloseArgs[64];
 int anhttpCloseArgsCount = 0;
-int anhttpCloseReturn = 0;
+int anhttpCloseReturn = (int)0;
 int anhttpClose(int fd) {
     if (anhttpCloseArgsCount == 64) assert(0);
     anhttpCloseArgs[anhttpCloseArgsCount].fd = fd;
@@ -41,7 +41,7 @@ int anhttpClose(int fd) {
 // anhttpAccept
 anhttpAcceptArgs_t anhttpAcceptArgs[64];
 int anhttpAcceptArgsCount = 0;
-int anhttpAcceptReturn = 0;
+int anhttpAcceptReturn = (int)0;
 int anhttpAccept(int fd, struct sockaddr *address, socklen_t *address_len) {
     if (anhttpAcceptArgsCount == 64) assert(0);
     anhttpAcceptArgs[anhttpAcceptArgsCount].fd = fd;
@@ -54,7 +54,7 @@ int anhttpAccept(int fd, struct sockaddr *address, socklen_t *address_len) {
 // anhttpBind
 anhttpBindArgs_t anhttpBindArgs[64];
 int anhttpBindArgsCount = 0;
-int anhttpBindReturn = 0;
+int anhttpBindReturn = (int)0;
 int anhttpBind(int socket, const struct sockaddr *address, socklen_t address_len) {
     if (anhttpBindArgsCount == 64) assert(0);
     anhttpBindArgs[anhttpBindArgsCount].socket = socket;
